requirements.txt
# requirements.txt
docker>=5.0.0
aiohttp>=3.8.0
prometheus_client>=0.12.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pymongo>=3.12.0
psutil>=5.8.0
paramiko>=2.8.0
pytest>=6.0.0
pytest-asyncio>=0.15.0
pytest-cov>=2.12.0
python-dotenv>=0.19.0